backend/requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.1
python-dotenv>=1.0
PyJWT>=2.8